Require GHC 8.0+
`hashable-1.4.*` adds an `Eq1` superclass to `Hashable1`, which means that in
order to define a `Hashable1 Vector` instance, we must ensure that an `Eq1
Vector` instance is always in scope. The `vector` library only provides this
instance for `base-4.9.*` (GHC 8.0.*) or later, so I have adjusted the version
bounds accordingly and dropped support for pre-8.0 GHCs. The alternative (using
CPP to conditionally backport the `Eq1 Vector` instance as an orphan instance)
sounds much less pleasant to maintain.